requirements.txt
qdrant-client[fastembed]
fastapi
uvicorn 
python-docx
openai
python-dotenv
sentence-transformers
python-multipart
openpyxl
PyMuPDF
requests
pathlib
pytz
openai==0.28
motor
groq

